[SofaSimulationCore] Move more MechanicalVisitors in their own files
